Move functions back into state_bits.py
