+++ src/open_prime_hunters_rando/patching/entities/shield_key_patches.py
@@ -4,55 +4,14 @@ from open_prime_hunters_rando.parsing.file_manager import FileManager
 from open_prime_hunters_rando.parsing.formats.entities.entity_types.artifact import Artifact
 from open_prime_hunters_rando.parsing.formats.entities.entity_types.camera_sequence import CameraSequence
 from open_prime_hunters_rando.parsing.formats.entities.entity_types.enemy_spawn import EnemySpawn
-from open_prime_hunters_rando.parsing.formats.entities.entity_types.item_spawn import ItemSpawn
 from open_prime_hunters_rando.parsing.formats.entities.entity_types.trigger_volume import (
     TriggerVolume,
     TriggerVolumeType,
 )
 from open_prime_hunters_rando.parsing.formats.entities.enum import Message
-from open_prime_hunters_rando.patching.entities.state_bits import STATE_BIT_SHIELD_KEY_MAPPING
 
 
-def create_shield_key_messages() -> list[str]:
-    pickup_messages = []
-    for key_data in STATE_BIT_SHIELD_KEY_MAPPING.values():
-        pickup_messages.append(
-            "PSHIELD KEY FOUND\\"
-            f"{key_data.unlock_message} in {key_data.area_name.upper()} - {key_data.room_name.upper()}!"
-        )
-    return pickup_messages
-
-
-def patch_shield_keys(file_manager: FileManager) -> None:
-    _create_shield_key_triggers(file_manager)
-    _patch_shield_key_rooms(file_manager)
-
-
-def _create_shield_key_triggers(file_manager: FileManager) -> None:
-    for state_bit, shield_key_data in STATE_BIT_SHIELD_KEY_MAPPING.items():
-        # Get the Shield Key
-        entity_file = file_manager.get_entity_file(shield_key_data.area_name, shield_key_data.room_name)
-        shield_key = entity_file.get_entity(shield_key_data.entity_id, ItemSpawn)
-
-        # Create a new trigger that checks if the state bit is set
-        # If set, it sends out the original message of the shield key
-        shield_key_trigger = TriggerVolume.create(
-            node_name=shield_key.node_name,
-            layer_state=shield_key.layer_state,
-            subtype=TriggerVolumeType.STATE_BITS,
-            required_state_bit=state_bit,
-            parent_id=shield_key.notify_entity_id,
-            parent_message=shield_key.collected_message,
-        )
-        entity_file.append_entity(shield_key_trigger)
-
-        # Sets the required state bit for the shield key
-        shield_key.notify_entity_id = -1
-        shield_key.collected_message = Message.SET_TRIGGER_STATE
-        shield_key.collected_message_param1 = state_bit
-
-
-def _patch_shield_key_rooms(file_manager: FileManager) -> None:
+def patch_shield_key_rooms(file_manager: FileManager) -> None:
     _high_ground(file_manager)
     _elder_passage(file_manager)
     _piston_cave(file_manager)
