+++ src/open_prime_hunters_rando/patching/text/string_tables_patches.py
@@ -2,7 +2,7 @@ from enum import Enum
 
 from open_prime_hunters_rando.parsing.file_manager import FileManager, Language
 from open_prime_hunters_rando.parsing.formats.string_tables import ScanCategory, ScanSpeed, StringTable
-from open_prime_hunters_rando.patching.entities.shield_key_patches import create_shield_key_messages
+from open_prime_hunters_rando.patching.entities.state_bits import create_shield_key_messages
 
 
 class StringTables(Enum):
